+++ CHANGELOG.md
@@ -12,10 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated everything to .NET 8
 - Enabled nullable reference types
 - Changed app builder initialization code to resemble the current default ASP.NET Core hosting model
+- If a default verb is defined, the verb passed by the user must strictly match one of the available verbs
 - Updated NLog to v5.3.3
 - Updated NLog.Extensions.Logging to v5.3.12
 
 ### Added
+
 - Commands are now async by default, they must use the new `IDotNetConsoleCommand` interface instead of `IConsoleAppCommand` and `IAsyncConsoleAppCommand`.
 
 ### Removed